Reorder comment nodes to go right before the node under them
Order of non-comment nodes are preserved. Each comment node is removed
from the containing nodes seq, and a position for it is determined by
looking at the coordinates for all other nodes. The comment node is
re-inserted right before the highest (but still under the comment node)
and leftmost node.

I also noticed that some nodes output null where they reference other
nodes. Notably, `mongodb` (config) nodes would be `null` when referenced
in `mongodb out` nodes. This was due to converting nodes to JS one tab
at a time, even though config nodes are not on a given tab. The
workaround was to pass along `all-nodes` for reference purposes.